refactor(queue): extract logger to shared utility and inject externally
- Move `createQueueWorkerLogger` to `src/shared/queue/logger.ts`
- Create per-worker `logger.ts` that configures via `createQueueWorkerLogger`
- Inject logger via callbacks instead of building it inside the handler
- Remove `sentryLogs` from `QueueWorkerConfig` (now owned by logger)
- Change envelope `type` from `z.literal` to `z.enum` for future workers
