+++ src/shared/queue/create-handler.ts
@@ -2,9 +2,8 @@ import * as Sentry from '@sentry/node';
 import type { SQSBatchResponse, SQSHandler } from 'aws-lambda';
 import { z } from 'zod';
 
-import { flattenForSentry } from '@/utils/flatten-for-sentry';
-
 import { messageEnvelopeSchema } from './envelope';
+import { QueueWorkerLogger } from './logger';
 import {
   checkIsProcessed,
   markProcessed,
@@ -14,57 +13,19 @@ import {
 export interface QueueWorkerConfig {
   name: string;
   maxReceiveCount: number;
-  sentryLogs: 'all' | 'error' | 'none';
   parsePayload: (payload: unknown) => unknown;
   dedupTtlMs?: number;
 }
 
-export interface QueueWorkerLogger {
-  info: (message: string, extras?: Record<string, unknown>) => void;
-  error: (message: string, extras?: Record<string, unknown>) => void;
-}
-
 export interface QueueWorkerCallbacks {
-  onProcess: (
-    logger: QueueWorkerLogger,
-    job: unknown,
-    messageId: string,
-  ) => Promise<void>;
-}
-
-function createLogger(name: string, sentryLogs: string) {
-  const component = `${name}-worker`;
-
-  function info(message: string, extras?: Record<string, unknown>): void {
-    console.info(`[${new Date().toISOString()}] ${message}`, extras ?? '');
-
-    if (sentryLogs === 'all') {
-      Sentry.logger.info(
-        `[${component}] ${message}`,
-        flattenForSentry(extras ?? {}),
-      );
-    }
-  }
-
-  function error(message: string, extras?: Record<string, unknown>): void {
-    console.error(`[${new Date().toISOString()}] ${message}`, extras ?? '');
-
-    if (sentryLogs === 'all' || sentryLogs === 'error') {
-      Sentry.logger.error(
-        `[${component}] ${message}`,
-        flattenForSentry(extras ?? {}),
-      );
-    }
-  }
-
-  return { info, error };
+  onProcess: (job: unknown, messageId: string) => Promise<void>;
+  logger: QueueWorkerLogger;
 }
 
 export function createQueueWorkerHandler(
-  config: QueueWorkerConfig,
-  { onProcess }: QueueWorkerCallbacks,
+  { maxReceiveCount, name, parsePayload, dedupTtlMs }: QueueWorkerConfig,
+  { onProcess, logger }: QueueWorkerCallbacks,
 ): SQSHandler {
-  const logger = createLogger(config.name, config.sentryLogs);
   const processedKeys: QueueProcessedKeys = new Map();
 
   return async (event): Promise<SQSBatchResponse> => {
@@ -80,11 +41,7 @@ export function createQueueWorkerHandler(
         const envelope = messageEnvelopeSchema.parse(body);
 
         if (
-          checkIsProcessed(
-            envelope.idempotencyKey,
-            processedKeys,
-            config.dedupTtlMs,
-          )
+          checkIsProcessed(envelope.idempotencyKey, processedKeys, dedupTtlMs)
         ) {
           logger.info('Duplicate message skipped', {
             messageId: record.messageId,
@@ -93,18 +50,16 @@ export function createQueueWorkerHandler(
           continue;
         }
 
-        const job = config.parsePayload(envelope.payload);
-        await onProcess(logger, job, record.messageId);
+        const job = parsePayload(envelope.payload);
+        await onProcess(job, record.messageId);
 
         markProcessed(envelope.idempotencyKey, processedKeys);
 
-        logger.info(`${config.name} processed`, {
-          messageId: record.messageId,
-        });
+        logger.info('Message processed', { messageId: record.messageId });
       } catch (err) {
         const receiveCount = Number(record.attributes.ApproximateReceiveCount);
 
-        logger.error(`${config.name} processing failed`, {
+        logger.error('Message processing failed', {
           messageId: record.messageId,
           receiveCount,
           error: err instanceof Error ? err.message : String(err),
@@ -113,12 +68,12 @@ export function createQueueWorkerHandler(
         if (
           err instanceof z.ZodError ||
           err instanceof SyntaxError ||
-          receiveCount === config.maxReceiveCount
+          receiveCount === maxReceiveCount
         ) {
           Sentry.captureException(err, {
             captureContext: {
               level: 'error',
-              tags: { worker: config.name },
+              tags: { worker: name },
               extra: { messageId: record.messageId, receiveCount },
             },
           });
