+++ src/workers/email/logger.ts
@@ -1,0 +1,5 @@
+import { createQueueWorkerLogger } from '@/shared/queue/logger';
+
+import { env } from './env';
+
+export const logger = createQueueWorkerLogger('email', env.SENTRY_LOGS);
